+++ CPP01/ex01/zombieHorde.cpp
@@ -12,7 +12,7 @@
 
 #include "Zombie.hpp"
 
-Zombie	*Zombie::zombieHorde(int N, std::string name)
+Zombie	*Zombie::zombieHorde(size_t N, std::string name)
 {
 	Zombie *zomb = new Zombie[N];
 	for(int i = 0; i < N; i++)
